0.0.1 -> 0.03
Updated the readme (done in 0.0.2)
Added more text in the mod settings panels
Added color coding to the main settings panel; panels get their background colors set to their variant colors, and the titles of the categories have a tint. This can be disabled in game
Organized code
Fixed some miscellaneous issues
Added the ability to change the color of the classic charge bar icon. By default it is the corresponding variation color